Move clients under da_service module
