+++ crates/adapters/celestia/src/da_service/client/mod.rs
@@ -1,0 +1,3 @@
+/// Celestia client implementations
+pub mod twinkle;
+pub mod vanilla;
