+++ crates/adapters/celestia/src/da_service/client/vanilla.rs
